# C#: Optimize parsing of some primitive and wrapper types