@jtattermusch 2019-11-04
I'll review the C# side, I'd like @gerben-s to check the correctness of the wrapper parsing optimizations.