@chrisdunelm 2019-11-05
The "TimerTest" benchmark shows the following improvement:
Benchmark on a Linux machine.

Before:
```
|            Method |    Mean |    Error |   StdDev |
|------------------ |--------:|---------:|---------:|
|      CsvBenchmark | 2.137 s | 0.0472 s | 0.0484 s |
| ProtobufBenchmark | 2.352 s | 0.0450 s | 0.0421 s |
```

After:
```
|            Method |    Mean |    Error |   StdDev |
|------------------ |--------:|---------:|---------:|
|      CsvBenchmark | 2.108 s | 0.0192 s | 0.0161 s |
| ProtobufBenchmark | 1.898 s | 0.0313 s | 0.0292 s |
```