@jtattermusch 2019-11-05
Merging (and will backport to v3.10.x)